requirements.txt
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-Table==0.5.0
Flask-WTF==0.14.3
getmac==0.8.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
keyboard==0.13.5
MarkupSafe==1.1.1
MouseInfo==0.1.3
Pillow==8.1.0
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
python3-xlib==0.15
PyTweening==1.0.3
pytz==2020.5
requests==2.25.1
SQLAlchemy==1.3.22
urllib3==1.26.3
Werkzeug==1.0.1
WMI==1.4.9
WTForms==2.3.3
